+++ src/app/components/CoinPageComponents/CoinDescription.tsx
@@ -2,9 +2,9 @@ const CoinDescription = ({ coinData, darkMode }) => {
    const description = coinData.description.en;
 
    return (
-      <div className="flex justify-center items-center ">
+      <div className="flex absolute md:relative justify-center items-center invisible md:visible">
          <div
-            className={`${darkMode ? "duration-300 text-white" : "duration-300 text-[#1e1932]"} w-[800px] text-sm`}
+            className={`duration-300 ${darkMode ? " text-white" : " text-[#1e1932]"} w-[800px] text-sm`}
             dangerouslySetInnerHTML={{ __html: description }}
          ></div>
       </div>
